Rename CommitmentFieldAccessTreeAnnot to InitializationFieldAccess..., move to top-level class, and improve comments
